캐릭터 정보창 구현

+++ Assets/Scripts/Monster/MonsterStatus.cs
@@ -7,7 +7,7 @@ public class MonsterStatus : MonoBehaviour
     [SerializeField] public string MonsterName; // 플레이어 이름
     [SerializeField] public float MonsterMaxHealth;
     [SerializeField] public float MonsterCurrentHealth;
-    [SerializeField] public float MonsterDamage = 10f;
+    [SerializeField] public int MonsterDamage = 10;
     [SerializeField] public int MonsterEXP = 1;
 
     void Start()
